+++ pygasflow/interactive/pages/normal_shock.py
@@ -19,8 +19,8 @@ class NormalShockSection(FlowSection):
                 filename="normal_shock",
                 columns_map={
                     "gamma": "gamma",
-                    "m1": "Upstream Mach",
-                    "m2": "Downstream Mach",
+                    "mu": "Upstream Mach",
+                    "md": "Downstream Mach",
                     "pr": "P2/P1",
                     "dr": "rho2/rho1",
                     "tr": "T2/T1",
@@ -29,7 +29,7 @@ class NormalShockSection(FlowSection):
             ),
         ])
         params.setdefault("wrap_in_card", False)
-        params.setdefault("input_parameter", "m1")
+        params.setdefault("input_parameter", "mu")
         super().__init__(**params)
         self.compute()
 
@@ -38,8 +38,8 @@ class NormalShockPage(FlowPage):
     input_parameter = param.Selector(
         label="Select parameter:",
         objects={
-            "Upstream Mach number": "m1",
-            "Downstream Mach number": "m2",
+            "Upstream Mach number": "mu",
+            "Downstream Mach number": "md",
             "Pressure Ratio P2/P1": "pressure",
             "Density Ratio rho2/rho1": "density",
             "Temperature Ratio T2/T1": "temperature",
@@ -63,9 +63,9 @@ class NormalShockPage(FlowPage):
     def _validate_input_value(self):
         # set appropriate default values so that no errors are raised
         # when user changes `input_parameter`
-        if self.input_parameter == "m1":
+        if self.input_parameter == "mu":
             self.input_value = "2"
-        elif self.input_parameter == "m2":
+        elif self.input_parameter == "md":
             self.input_value = "0.5773502691896257"
         elif self.input_parameter == "pressure":
             self.input_value = "4.5"